requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-dropbox-storage==0.4.1
django-js-asset==1.2.2
django-storages==1.9.1
djangorestframework==3.11.0
dropbox==10.2.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
openpyxl==3.0.4
psycopg2-binary==2.8.4
pylint==2.5.3
python-decouple==3.3
pytz==2019.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.27
wfastcgi==3.0.0
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
